+++ CHANGELOG.md
@@ -18,6 +18,7 @@ If you haven't upgraded to v6.0.0 yet, please read v6.0.0 changelog first.
 * ConverseJS: Don't send state notifications. #134
 * ConverseJS: Don't show join/leave messages. #134
 * ConverseJS: remove browser notifications, and browser tab message count.
+* ConverseJS: in readonly mode, don't show any info messages (except when disconnected).
 
 ### Minor changes and fixes